<refactor>(project): 피드백 내용 적용
