+++ src/main/java/gift/controller/WishController.java
@@ -5,9 +5,8 @@ import gift.dto.request.WishAddRequestDto;
 import gift.dto.request.WishDeleteRequestDto;
 import gift.dto.request.WishUpdateRequestDto;
 import gift.dto.response.WishIdResponseDto;
-import gift.entity.User;
 import gift.entity.WishProduct;
-import gift.service.WishService;
+import gift.service.WishServiceImpl;
 import gift.wishPreProcess.LoginMember;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -26,46 +25,46 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/wishes")
 public class WishController {
 
-    private final WishService wishService;
+    private final WishServiceImpl wishServiceImpl;
 
-    public WishController(WishService wishService) {
-        this.wishService = wishService;
+    public WishController(WishServiceImpl wishServiceImpl) {
+        this.wishServiceImpl = wishServiceImpl;
     }
 
     @PostMapping("")
     public ResponseEntity<WishIdResponseDto> addToWish(
         @RequestBody @Valid WishAddRequestDto wishAddRequestDto,
-        @LoginMember User currentUser
+        @LoginMember String userEmail
     ) {
         return ResponseEntity.status(HttpStatus.CREATED)
-            .body(wishService.addProduct(wishAddRequestDto, currentUser.email()));
+            .body(wishServiceImpl.addProduct(wishAddRequestDto, userEmail));
     }
 
     @GetMapping("")
     public ResponseEntity<List<WishProduct>> getWishItem(
-        @LoginMember User currentUser
+        @LoginMember String userEmail
     ) {
         return ResponseEntity.status(HttpStatus.OK)
-            .body(wishService.getWishList(currentUser.email()));
+            .body(wishServiceImpl.getWishList(userEmail));
     }
 
     @DeleteMapping("/{wishId}")
     public ResponseEntity<Void> deleteWish(
         @RequestBody @Valid WishDeleteRequestDto productName,
-        @LoginMember User currentUser,
+        @LoginMember String userEmail,
         @PathVariable Long wishId) {
 
-        wishService.deleteProduct(currentUser.email(), wishId, productName);
+        wishServiceImpl.deleteProduct(userEmail, wishId, productName);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     @PatchMapping("/{wishId}")
     public ResponseEntity<Void> updateWish(
         @RequestBody @Valid WishUpdateRequestDto wishUpdateRequestDto,
-        @LoginMember User currentUser,
+        @LoginMember String userEmail,
         @PathVariable Long wishId
     ) {
-        wishService.updateProduct(wishId, currentUser.email(), wishUpdateRequestDto);
+        wishServiceImpl.updateProduct(wishId, userEmail, wishUpdateRequestDto);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
