+++ src/main/java/gift/wishPreProcess/LoginMemberArgumentResolver.java
@@ -1,11 +1,8 @@
 package gift.wishPreProcess;
 
 import gift.auth.JwtTokenHandler;
-import gift.entity.User;
-import gift.exception.UserNotFoundException;
-import gift.service.UserService;
+import gift.service.UserServiceImpl;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.Optional;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -17,30 +14,30 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
 
     private final JwtTokenHandler jwtTokenHandler;
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
-    public LoginMemberArgumentResolver(JwtTokenHandler jwtTokenHandler, UserService userService) {
+    public LoginMemberArgumentResolver(JwtTokenHandler jwtTokenHandler,
+        UserServiceImpl userServiceImpl) {
         this.jwtTokenHandler = jwtTokenHandler;
-        this.userService = userService;
+        this.userServiceImpl = userServiceImpl;
     }
 
     @Override
     public Object resolveArgument(MethodParameter parameter,
         ModelAndViewContainer mavContainer,
         NativeWebRequest webRequest,
-        WebDataBinderFactory binderFactory) throws Exception {
+        WebDataBinderFactory binderFactory) {
 
         HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
 
-        Optional<User> user = userService.getUserByEmail(
-            jwtTokenHandler.getEmailFromHeader(
-                request.getHeader("Authorization")));
-        return user.orElseThrow(() -> new UserNotFoundException("사용자를 찾을 수 없습니다"));
+        String authHeader = request.getHeader("Authorization");
+        authHeader = authHeader.substring(7);
+        return jwtTokenHandler.getEmailFromToken(authHeader);
     }
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
         return parameter.hasParameterAnnotation(LoginMember.class)
-            && parameter.getParameterType().equals(User.class);
+            && parameter.getParameterType().equals(String.class);
     }
 }
